refactor(jobsdb): use a request executor to encapsulate execution mode (concurrent or queued)
In the previous implementation, each method that wanted to perform a read/write operation was repeating the same logic:

Check if read/write should be done through a channel or not.
If not, execute the function immediately.
If a channel should be used, prepare another struct holding the request type, parameters & response channel, push this struct to the appropriate channel, the worker then would read the request, execute a different function depending on the request type and publish the response to the response channel.
Now instead, we always create a command that holds the function to be executed and this is sent to an executor.
The executor, depending on the runtime configuration of the system will either execute the function synchronously, or push it to the relevant channel and wait until a response is received.

A few advantages introduced with this refactoring:

DRY
No need to copy function parameters around
The executor doesn't need to know any details about the function being called